♻️ refactor: PostPage 마크업에 따른  공통컴포넌트 수정 (#67)
